Nettopreise Eingabe von ganzzahligem Nettopreis und ausgabe von 10 Nettopreisen im Abstand von 1

+++ Aufgabe 1/Taschenrechner.c
@@ -21,7 +21,7 @@ int minus() {
     return a - b;
 }
 
-float divide() {
+int divide() {
     int a, b = 0;
     printf("Enter first number:\n");
     scanf("%d", &a);
